Updated the ESP  Serial Board code to be a single sketch
I updated the code to allow all ESP Serial boards to run the same sketch.  You would need to comment out other boards in lines 18-21.

This update also enables the Access Point on the Body Controller to get rid of the external router.   This is initially working but I haven't stress tested it yet for reliability. or range or performance .